update 컨트롤러 및 서비시에 인터페이스 적용 및 불필요한 파일 삭제
