+++ src/app.controller.ts
@@ -2,9 +2,21 @@ import { Controller, Get, HttpStatus, Param, Res } from "@nestjs/common";
 import { Response } from "express";
 import { AppService } from "./app.service";
 import { version } from "../package.json";
+import { IGlasses, ITodos } from "./common/common.interface";
+import { IServiceEndpoint } from "./common/common.endpoint.interface";
 @Controller()
-export class AppController {
+export class AppController implements IServiceEndpoint {
   constructor(private readonly appService: AppService) {}
+
+  @Get(":userId/today-todos")
+  getTodayTodos(@Param("userId") userId: string): Promise<ITodos> {
+    return this.appService.getTodayTodos(userId);
+  }
+
+  @Get(":userId/glass")
+  getPlantingGlass(@Param("userId") userId: string): Promise<IGlasses> {
+    return this.appService.getPlantingGlass(userId);
+  }
 
   @Get("ping")
   sendStatusOK(@Res() res: Response): void {
@@ -14,15 +26,5 @@ export class AppController {
   @Get("version")
   getVersion() {
     return version;
-  }
-
-  @Get(":userId/glass")
-  getGlasses(@Param() param) {
-    return this.appService.getPlantingGlass(param.userId);
-  }
-
-  @Get(":userId/today-todos")
-  getTodayTodos(@Param() param) {
-    return this.appService.getTodayTodos(param.userId);
   }
 }
