+++ src/app.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from "@nestjs/common";
+import { IServiceEndpoint } from "./common/common.endpoint.interface";
 import { randomColor, randomId, randomInt0to10 } from "./common/common.function";
-import { IGlass, IGlasses, ITodos } from "./common/common.interface copy";
-import { IServiceEndpoing } from "./common/common.service-interface";
+import { IGlass, IGlasses, ITodos } from "./common/common.interface";
 @Injectable()
-export class AppService implements IServiceEndpoing {
+export class AppService implements IServiceEndpoint {
   async getPlantingGlass(userId: string): Promise<IGlasses> {
     console.log(`getGlasses : ${userId}`);
     const result: IGlass[] = [];
